Return account storage string
