+++ blockchain/flowkit.go
@@ -353,55 +353,18 @@ func (fk *flowKit) createAccount() (*flow.Account, error) {
 }
 
 func (fk *flowKit) getAccountStorage(address flow.Address) (string, error) {
-	const StorageIteration = `
-	pub fun main(address: Address) : AnyStruct{
-	
-		var res :  [{String:AnyStruct}] = []
-	
-		getAuthAccount(address).forEachStored(fun (path: StoragePath, type: Type): Bool {
-			res.append(
-			{
-				"path" : path,
-				"type" : type.identifier,
-				"value":  type.isSubtype(of: Type<AnyStruct>()) ?
-								getAuthAccount(address).borrow<&AnyStruct>(from: path)! as AnyStruct
-								: getAuthAccount(address).borrow<&AnyResource>(from: path)! as AnyStruct
-			})
-			return true
-		})
-	
-		getAuthAccount(address).forEachPublic(fun (path: PublicPath, type: Type): Bool {
-			res.append(
-			{
-				"path" : path,
-				"type" : type.identifier,
-				"value":  getAuthAccount(address).getLinkTarget(path)
-			})
-			return true
-		})
-	
-		getAuthAccount(address).forEachPrivate(fun (path: PrivatePath, type: Type): Bool {
-			res.append(
-			{
-				"path" : path,
-				"type" : type.identifier,
-				"value":  getAuthAccount(address).getLinkTarget(path)
-			})
-			return true
-		})
-		return res
-	}`
-
 	args := []string{fmt.Sprintf(`{"type":"Address","value":"0x%s"}`, address.Hex())}
 	val, _, err := fk.executeScript(StorageIteration, args)
 	if err != nil {
 		return "", err
 	}
 
+	/* TODO: Parse account storage
 	_, err = ParseAccountStorage(val)
 	if err != nil {
 		return "", err
 	}
+	*/
 
 	return val.String(), nil
 }
