+++ e2eTest/account_test.go
@@ -80,7 +80,7 @@ func TestAccountDeployedContracts(t *testing.T) {
 }
 
 func TestAccountStorage(t *testing.T) {
-	t.Skip("Account storage not implemented with flowKit") // TODO: Add back account storage
+	t.Skip("Account storage parsing not implemented with FlowKit") // TODO: Add back account storage
 
 	c := newClient()
